新加支付、完善产品和订单模块

+++ app/product/routes.py
@@ -8,9 +8,10 @@ from app.models import Product
 
 
 @bp.route('/add', methods=['GET'])
+@utils.check_admin
 def add():
     name = request.args.get('name', '')
-    price = request.args.get('price', 0, int)
+    price = request.args.get('price', 0, float)
     inventory = request.args.get('inventory', 0, int)
     if name == '':
         return utils.ret_err(-1, 'name is requested')
@@ -27,9 +28,13 @@ def add():
     return utils.ret_msg_objs('Suceess', product)
 
 
-@bp.route('/delete/<int:product_id>', methods=['GET'])
-def delete(product_id):
-    product = Product.query.filter_by(id=product_id).first()
+@bp.route('/delete', methods=['GET'])
+@utils.check_admin
+def delete():
+    p_id = request.args.get('p_id')
+    if p_id is None or p_id == '':
+        return utils.ret_err(-1, "p_id is required")
+    product = Product.query.filter_by(p_id=p_id).first()
     if product is None:
         return utils.ret_err(-1, "Product doesn't exists")
     db.session.delete(product)
@@ -37,9 +42,13 @@ def delete(product_id):
     return utils.ret_msg('Success')
 
 
-@bp.route('/update/<int:product_id>', methods=['GET'])
-def update(product_id):
-    product = Product.query.filter_by(id=product_id).first()
+@bp.route('/update', methods=['GET'])
+@utils.check_admin
+def update():
+    p_id = request.args.get('p_id')
+    if p_id is None or p_id == '':
+        return utils.ret_err(-1, "p_id is required")
+    product = Product.query.filter_by(p_id=p_id).first()
     if product is None:
         return utils.ret_err(-1, "Product doesn't exists")
     params = {}
@@ -60,26 +69,36 @@ def update(product_id):
         params["detail"] = detail
 
     if bool(params):
-        pro = Product.query.filter_by(id=product_id).update(params)
-        return utils.ret_msg_objs('Success', pro)
+        Product.query.filter_by(p_id=p_id).update(params)
+        db.session.commit()
+        return utils.ret_msg_objs('Success', product)
     return utils.ret_msg('Nothing happen')
 
 
-@bp.route('/list/<int:product_id>', methods=['GET'])
-def list_product(product_id):
-    product = Product.query.filter_by(id=product_id).first()
+@bp.route('/list', methods=['GET'])
+@utils.check_admin
+def list_product():
+    p_id = request.args.get('p_id')
+    if p_id is None or p_id == '':
+        return utils.ret_err(-1, "p_id is required")
+    product = Product.query.filter_by(p_id=p_id).first()
     return utils.ret_objs(product)
 
 
-@bp.route('/list', methods=['GET'])
+@bp.route('/listAll', methods=['GET'])
+@utils.check_admin
 def list_all_product():
     products = Product.query.all()
     return utils.ret_objs(products)
 
 
-@bp.route('/<int:product_id>/listOrder', methods=['GET'])
-def list_order_by_product_id(product_id):
-    product = Product.query.filter_by(id=product_id).first()
+@bp.route('/listOrder', methods=['GET'])
+@utils.check_admin
+def list_order_by_product_id():
+    p_id = request.args.get('p_id')
+    if p_id is None or p_id == '':
+        return utils.ret_err(-1, "p_id is required")
+    product = Product.query.filter_by(p_id=p_id).first()
     if product is not None:
         return utils.ret_objs(product.orders)
     return utils.ret_err(-1, "Product doesn't exists")
